requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.2
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
executing==2.0.1
gunicorn==22.0.0
idna==3.7
ipython==8.24.0
jedi==0.19.1
matplotlib-inline==0.1.7
packaging==24.0
parso==0.8.4
prompt_toolkit==3.0.45
pure-eval==0.2.2
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.2
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
whitenoise==6.6.0
